// types/note.ts
export type NoteTag = 'Todo' | 'Work' | 'Personal' | 'Meeting' | 'Shopping';
export interface Note {
  id: string;
  title: string;
  content: string;
  tag: NoteTag;
  createdAt: string;
  updatedAt: string;
}

export interface FormValues {
  title: string;
  content: string;
  tag: NoteTag;
}
export interface CreateNote {
  title: string;
  content: string;
  tag: NoteTag;
}